requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.9

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Validation
pydantic==2.9.0
pydantic-settings==2.5.2
email-validator==2.2.0

# HTTP requests
requests==2.32.3
httpx==0.27.2